plugin: rename sopel.module into sopel.plugin
